// frontend/src/components/ui/PageHeader.d.ts
import React from 'react';
export interface PageHeaderProps {
    title: string;
    description?: string;
    subtitle?: string;
    icon?: React.ReactNode;
    action?: React.ReactNode;
    badge?: React.ReactNode;
}
export declare const PageHeader: React.FC<PageHeaderProps>;
//# sourceMappingURL=PageHeader.d.ts.map